new version
Delete button handler added

only New button handler to be implemented

Testing/usage :: if you delete element is not directly removed from the ListBox, not even when choosing another tab and returning. 
If  you close the Edit Box and reopen the item shows removed
A save and reload should work to permannently remove the MinMaxContractLine but i havent tested that.